requirements.txt
pandas
openpyxl
python-docx
docx2pdf
Pillow
reportlab
ttkbootstrap
tkinterdnd2